[refactoring] remove inline style
