+++ components/Footer.js
@@ -1,6 +1,6 @@
 import React from "react";
 import Image from 'next/image'
-import styles from '../styles/Footer.module.css'
+import styles from '../styles/components/Footer.module.css'
 
 export default function Footer() {
     return (
