+++ styles/pages/Battle.module.css
@@ -1,0 +1,22 @@
+.container {
+    width: 100%;
+    display: flex;
+    flex-wrap: wrap;
+    align-items: center;
+    justify-content: center;   
+}
+.title {
+    font-weight : 300;
+}
+
+.btn {
+    margin-left: 5px;
+    width: 300px;
+    height: 3rem;
+    background-color: white;
+    border: none;
+    font-family: 'Helvetica Neue, sans-serif';
+    font-weight: 700;
+    font-size:30px;
+    cursor: pointer;
+}
